Split sled_storage.rs into multiple files,
Store & StoreMut implementations locate in sled_storage/store.rs
AlterTable implementation locates in sled_storage/alter_table.rs
